feat: Initial limit pushdown in the scan (#3483)
Figured that `SELECT * FROM tbl LIMIT 10` is a pretty common use case
for data exploration and is probably worth including. I think this
closes #2593 and is also part of #2254.

Supporting limit with filters with our heavy use of `async move` seems
much harder because the limit has to be shared in some way. Doing that
for the stream (like when `into_stream`) is trivial with a
`StreamExt::scan` but for the current return type of
`ScanBuilder::build` that will require shared state and I'm not even
sure that would be correct given the ordering isn't guaranteed.

---------

Signed-off-by: Adam Gutglick <adam@spiraldb.com>