chore: fix type bound on Scalar::binary (#3930)
There isn't anything that impls `Into<Arc<ByteBuffer>>` except for
ByteBuffer, so the bound currently is kind of pointless. I've relaxed
the bound so that we get all of `Into<ByteBuffer>`, which covers a wider
range of types.

The internal Arc is just there to cut down on the size of
InnerScalarValue, there aren't any code paths where we actually get an
`Arc<ByteBuffer>` out of something and want to turn it into a scalar
directly.

Signed-off-by: Andrew Duffy <andrew@a10y.dev>