fix: remove optimize Op, replace with compact_buffers function (#3931)
Fix both fuzzers.

The first is failing b/c we're not running it with nightly toolchain.

The second is failing due to some issues with how Optimize operation is
implemented for RunEndArray. Currently Optimize for sliced types
subtracts out the offset, but that actually forces a decompression,
which is triggering our debug assertion that calling `optimize` never
makes the array smaller.

After chatting with @gatesn , think for now we should move compacting to
just be part of `impl Canonical`, as the main use-case is to call right
before compression. Removing it now lets us keep the API surface clean
while we stew on the right way to introduce this more widely (if at
all).

Tested with the fuzzer artifact that were failing before this change:

<img width="1508" height="696" alt="image"
src="https://github.com/user-attachments/assets/33afda84-6dcd-4cb5-8ae5-c4e7b3ef1003"
/>

---------

Signed-off-by: Andrew Duffy <andrew@a10y.dev>