varbinview zip kernel (#4054)
Uses the buffer deduplicating builder to construct the zipped array.
This guards against the pathological case where we are zipping two
varbinview arrays with a mask that has lots of contiguous slices. Each
`builder.extend_from_array(input.slice(..))` would duplicate the entire
buffers of `input`, and each slice in the mask would add the same
buffers to the result array over and over again.

---------

Signed-off-by: Onur Satici <onur@spiraldb.com>