refactor the LayoutStrategy (#4149)
Breaking this out as it was some refactoring I was already doing as part
of #3810

We are already passing around BoxFuture with our `SendableLayoutFuture`,
using the async_trait stuff just makes it more ergonomic.

The other thing I did is that I've updated our existing `LayoutStrategy`
impls to be parameterized by the children. This actually is completely
fine for us and doesn't harm extensibility. I also find it a bit clearer
too, and I think there is also a benefit of fewer unnecessary layers of
indirection.

I added more doc comments, and also put `VortexLayoutStrategy` behind a
builder, which unifies the compressor configuration.

Signed-off-by: Andrew Duffy <andrew@a10y.dev>