Inline the logic of the filter layout reader (#4089)
Follow up work is to store expressions as `Vec<Vec<Expr>>` (including
the projection expression as the last elemet). Then if an expression
ever hits 0% selectivity, we extract it into its own round trip, e.g.
`select C where A == 10 and B > 2` would become `[[A == 10], [B > 2,
C]]` if `A == 10` proves to be very selective.

This allows us to delay I/O for columns B and C until we have evaluated
over A.

---------

Signed-off-by: Nicholas Gates <nick@nickgates.com>