Feature: Add `FixedSizeListBuilder` (#4428)
Adds support for `FixedSizeListBuilder: ArrayBuilder`. I don't actually
_use_ the builder where it needs to be yet, but I am going to leave that
for a followup PR.

Additionally adds unit tests for the builder, making sure to test the
degenerate cases.

Also note that some of the TODOs with respect to the `append_zero` and
`append_null`s will be solved in the next PR I do so don't worry too
much about those...

---------

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>