Chore: Validate `Scalar::new` for null and clean up `DecimalScalar` (#4435)
Progress towards #4378 

~~I don't think that it makes a lot of sense to add `new_unchecked`
given that the only check we really need is `!dtype.is_nullable()`
(which is cheap), and by making the fields private we can guarantee this
invariant.~~

Edit: In the future the `Scalar::new` should validate that the dtype
matches the value that is passed in.

Also cleans up the `decimal` module in `vortex-scalar`.

---------

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>