Chore: Make all `ArrayBuilder`s consistent (#4450)
This PR refactors array builders for consistent APIs (and mostly
consistent implementations).

## Changes

This is really all of the changes from the other 4 PRs I made right
before this...

- `append_value` -> type-specific names (`append_bool`,
`append_decimal`, `append_list`)
- All builders now validate dtype compatibility in `extend_from_array`
- `is_superset_of`/`is_subset_of` for type checking
- Standardized method signatures across all builders
- Extracted common constants, improved docs
- Added `finish_into_specific` for all of the builders (this opens the
door to implementing `ToCanonical` for every `ArrayBuilder`, avoiding an
allocation)

Note that I avoided changing `VarBinViewBuilder` since there's a lot of
code in there that I don't understand.

Also note that there are a lot of breaking changes, but to be honest, a
bunch of `append_value`s that took completely different parameters is
too confusing.

---------

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>