Change evaluations to take MaskFuture (#4482)
We want the option to be able to run the conjunct and projection
expressions in parallel, at least in order to race their I/O.

Without a MaskFuture, we cannot begin to process the projection
evaluation without first completing the download and evaluation of the
conjuncts. We currently use pre-fetching to get around this, but it's a
little crude and we'd rather be explicit.

This PR does not race conjuncts, simply changes the signature.

---------

Signed-off-by: Nicholas Gates <nick@nickgates.com>