fix[vortex-array]: don't use input list field type in preferred conversion (#4481)
This would cause a panic in arrow when attempting to create a list when
the preferred conversion resulted in a different element array type due
to the different types.

This commit also simplifies the input arguments so that preferred type
conversion never has access to the canonical arrow type.

Ran into this while testing #4464

Signed-off-by: Alfonso Subiotto Marques <alfonso.subiotto@polarsignals.com>