Convert handles to hold weak reference to their runtime (#4624)
This fixes a problem where we can spawn a task that holds onto handle
(since it has a static lifetime, ergh) can keep a runtime alive even
after the caller drops it. For futures that use channels, e.g. the
FileRead I/O, this results in read futures that hang forever when the
runtime is dropped (because the executor holds the pending recv end of
the channel, and the executor is kept alive by the send end of the
channel...)

We also impl BlockingRuntime for CurrentThreadRuntime as this will
likely become the default runtime for language bindings to use.

---------

Signed-off-by: Nicholas Gates <nick@nickgates.com>