Move VortexRead over to new File I/O (#4626)
This PR moves our I/O layer over from the old pre-fetching segment
source to the new coalescing File I/O implementation on the Vortex
runtime.

Note: this PR currently includes #4624 

## Breaking Changes

* `VortexOpenOptions` is no longer generic.
* Fix: find/replace `VortexOpenOptions::file` ->
`VortexOpenOptions::new`
* `VortexOpenOptions::in_memory().open(...)` is replaced by
`VortexOpenOptions::new().open_buffer(...)`

## Performance

Generally this PR is about even, or slightly improved. There is some
variance on some of the S3 benchmarks.

The big regressions are in DuckDB on S3, where our read path has a
custom work-stealing queue that doesn't buffer enough of the scan to
perform effective pre-fetching. DataFusion doesn't have this problem, so
we see much less of an impact.

I want to follow up to this PR to switch over DuckDB to avoid this
getting large, and accept the short-term regression.

---------

Signed-off-by: Nicholas Gates <nick@nickgates.com>