Fix: `FixedSizeList` DuckDB exporter bug (#4610)
Tracking Issue: https://github.com/vortex-data/vortex/issues/4372

Changes checked out from:
https://github.com/vortex-data/vortex/pull/4605

Fixes a bug in the DuckDB exporter where if we do not flatten /
materialize the inner child of the fixed-size list before exporting the
vector to DuckDB, DuckDB will segfault after running a function like
`scan`.

Also fixes a bug where I didn't correctly append garbage values when the
validity is all null (I took that from the list implementation, which
doesn't need to append garbage values since null can be an empty list
there). I don't actually know if this is an issue to be honest, but it
seems wrong?

I was not able to figure out how to write regression tests for both of
these. The early return might not actually be an issue, but the
flattening at the end definitely is. However, I was not able to create a
minimally reproducible test for that, and it only fails in query 7 and 8
in `statpopgen.sql`.

```sql
-- 7. Read just one variant (this is the sixth one).
SELECT *
  FROM statpopgen
 WHERE "CHROM" == 'chr21'
   AND "POS" == 5030278;
-- 8. Read a 700 base-pair window of variants.
SELECT *
  FROM statpopgen
 WHERE "CHROM" == 'chr21'
   AND "POS" >= 5030300
   AND "POS" <= 5031000;
```

If anyone has an idea of how to recreate this easily that would be
appreciated!

Finally, I simplified the tests I wrote from before since a lot were
redundant.

---------

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>