Feature: `ListView` Compute Functions (#4712)
Tracking Issue: https://github.com/vortex-data/vortex/issues/4699

Adds compute functions to the `ListView` encoding.

Additionally refactors the the `list` `filter` implementation to be a
bit more succinct.

Also adds tests, and refactors some of the fixed-size list tests to use
more conformance tests.

Unresolved Questions:

~~How much do we care about carrying around extra data? Both `filter`
and `cast` can be implemented by pushing down the operation directly to
the `offsets` and `sizes` child arrays. This can leave quite a lot of
data in the `elements` array.~~

~~There are many things we can do here. In the second commit I trim the
ends of the `elements` array using the min and max indexes into the
`elements` (derived from computing `min/max` of `offsets` and `offsets +
sizes`). This could be more expensive than we want.~~

~~What we could do instead is simply filter/take `offsets`/`sizes` and
then allow the caller to compact / "clean up" by passing the final array
back through a builder (which can have different strategies to rebuild
the array). I think I am in favor of this (though I'm leaving the code
in the commit history for now so people can see it).~~

---------

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>