varbinview compact to zero copy whenever possible (#4827)
This also adds a way to the varbinview builder to compact arrays as they
get added to the builder.

The way this works is to calculate the overall utilisation (num bytes
referred from views per buffer), and keep buffers from the array to be
added if they already are highly utilised.
Another optimisation this brings is to track the used ranges per buffer,
and check if the byte usage within that range is above the utilisation
threshold, and slice the buffer (zero copy) during compaction. This is
useful when the varbinview array is constructed from our builders and
then sliced, meaning some buffer has half of its bytes unreferenced, we
take the other half without copying in those cases now.

---------

Signed-off-by: Onur Satici <onur@spiraldb.com>