Chore: Clean up benchmarks website (#4921)
Because lance is so much slower in some of the queries than every other
target, it makes a lot of charts harder to see everything.

EDIT: They are actually not showing up at all on the TPCH

EDIT AGAGIN: I'm going to use this PR to fix the two other problems. The
first is that because we do not run lance on every PR the summaries will
disappear. The second is that TPC-H queries are not showing up.

---------

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>