feat: GithubArchive benchmark for nested data (#4960)
GithubArchive is an archival dataset of Github event logs, hosted at
https://www.gharchive.org/.

It's also distributed as part of RealNest from CWI.

This is our first benchmark to test nested data access with Vortex.


I want to get this in before we merge #4942 or any other follow on work
for nested data support.

TODOs:

- [x] Add more data files. Currently queries finish way too fast
- [x] Add a few more queries

---------

Signed-off-by: Andrew Duffy <andrew@a10y.dev>