Feature: Add `vortex-vector` (#5010)
Tracking Issue: https://github.com/vortex-data/vortex/issues/4492

Continuation of https://github.com/vortex-data/vortex/pull/4954

This is mostly boilerplate code with a bit more documentation than
normal.

There are still a few bugs in the implementation (documented by TODOs),
but I would like this PR to just focus on the structure of the crate
rather than the logic.

Also add logic in other crates that was missing before.

---

The general structure of the crate is as such:

```
vortex-vector
├── Cargo.toml
└── src
    ├── <type>
    │   ├── mod.rs
    │   ├── vector.rs
    │   └── vector_mut.rs
    ├── lib.rs
    ├── macros.rs
    ├── ops.rs
    ├── private.rs
    ├── vector.rs
    └── vector_mut.rs
```

The exceptions to this structure will be `primitive`, which needs a
generic type as well (so there will be extra files), a `varbin` module
that holds both `binary` and `string` modules, and probably the 2 list
types might share a lot of code (we'll see once I get there).

Again, a lot of the logic here is not super clean and some of it is
wrong, but I'd like to get this merged first before I fix those.

---------

Signed-off-by: Nicholas Gates <nick@nickgates.com>
Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>
Co-authored-by: Nicholas Gates <nick@nickgates.com>