Feature: Struct Vectors (#5094)
Tracking Issue: https://github.com/vortex-data/vortex/issues/5028

Adds `StructVector` and `StructVectorMut`.

Note: It would be really nice to have something like serde values that
we can iterate and take as an iterator to construct these struct
vectors, otherwise we are left with what I currently have on this PR,
which imo is very unwieldy. This can come in a later PR though.

The second commit fixes some errors and also removes all of the `new`
functions for vectors (as per an offline discussion with @gatesn).

---------

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>