feat[gpu]: use BufferHandle to the array vtable build (#5565)
We want to have array which are backed by either CPU or GPU/device
buffers.

The first stage is to create that enum and change the build array
method.
Then add this to the segment source and array construction.
Then we can go through each array and support non-cpu buffers.


There is an open question if we want to have this logical-scan-plann
have buffers or BufferIDs

---------

Signed-off-by: Joe Isaacs <joe.isaacs@live.co.uk>