Feature: add `take` for primitive and bool vectors (#5540)
Adds `take` implementations and tests for:

- `BitBuffer`
- `Mask`
- `BoolVector`
- `PrimitiveVector`

For `BitBuffer`, I don't think it makes sense to implement take with
nullable indices since I would argue that there isn't a clear behavior
to have (because it might not be obvious that `true` means not null and
`false` means null). But for `Mask` it feels a bit different, we are
using it for validity in all of our vectors, so I think it makes a bit
more sense (and it will make the implementation of `take` for the rest
of the vectors easier).

We represent indices in 3 different ways here (a generic indices slice
of unsigned integers, generic unsigned pvector, and primitive vector). I
feel quite strongly that we should not allow signed indices, as that
opens the door to too many bugs.

This is also based on top of
https://github.com/vortex-data/vortex/pull/5537, so that needs to merge
first.

---------

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>