feature: add `__version__` attribute to Python package (#5641)
This PR adds the `__version__` attribute to the Python package. This is
commonly added to Python packages to help observe the version of a
package through the package API. I tried to make this with dynamic
versioning in mind, but please don't hesitate to provide feedback if
this misses the mark.

Closes #5615

---------

Signed-off-by: d33bs <ekgto445@gmail.com>
Signed-off-by: Dave Bunten <ekgto445@gmail.com>