Benchmarks re-org (#5676)
This PR makes two big changes to our benchmarking setup:
1. It splits duckdb, datafusion and lance benchmarks into their own
crates, so they don't depend on each other, minimizing the friction when
bumping dependencies like Arrow or DataFusion.
2. Introduces a new benchmarking tool - `vx-bench` which allows
comparing various runs, persisting results and takes care of building
and running the individual binaries. `vx-bench` is written in Python
using `typer`.

I've tried to not touch the CI setup too much, but it can be simplified
and improved significantly to fit this structure.

## Notes for reviewers
1. DF/Arrow results are worse because the data is chunked differently,
we use an `MemTable` which is just a `Vec<RecordBatch>`, which is now
longer.
2. ~I'm looking into the differences in Vortex results, not sure what's
going on there.~ Turns out the `vortex-compact` results are slower
because they actually run on data compressed with the compact
compressor! Some of the tables are about half the size, but they are
slower to decompress/process.
4. ~Seems like we've been dropping some compression benchmark results,
so they dont have a baseline (or I changed the name) I'm trying to
figure out what's going on there.~ I've changed the name of some
compression benchmarks, ratios have the same name though.

## `vx-bench` demo


![render1765993255912](https://github.com/user-attachments/assets/0e0d3c9d-e7f9-47b6-bdff-e41a34e46348)

---------

Signed-off-by: Adam Gutglick <adam@spiraldb.com>